lighten SearchTreeRao and LinearRao packages
