+++ ra-optimisation/linear-rao/src/main/java/com/farao_community/farao/linear_rao/JsonLinearRaoResult.java
@@ -5,10 +5,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.farao_community.farao.linear_rao.json;
+package com.farao_community.farao.linear_rao;
 
 import com.farao_community.farao.commons.FaraoException;
-import com.farao_community.farao.linear_rao.LinearRaoResult;
 import com.farao_community.farao.linear_rao.LinearRaoResult.SystematicSensitivityAnalysisStatus;
 import com.farao_community.farao.rao_api.json.JsonRaoResult;
 import com.fasterxml.jackson.core.JsonGenerator;
