+++ ra-optimisation/search-tree-rao/src/test/java/com.farao_community.farao.search_tree_rao/SearchTreeRaoParametersTest.java
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.search_tree_rao.config;
+package com.farao_community.farao.search_tree_rao;
 
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.powsybl.commons.config.PlatformConfig;
